Add a linked list datatype
I've realised after much too long that we are probably best off
storing the children of a tree inside a linked list instead of
trying to mind-numbingly reallocate the same spot of memory and
set the last space of that memory to NULL. Will it be slower?
Probably. Will I eventually get this finished? Yes.