Fix bad #include management
Was becomming a huge hassle to deal with includes in every file,
so I decided to just have every include in eval.h. If there are any
problems with other files not getting the right types/functions at
the right times, those functions need to be defined without bodies
at the top of those files.